lib: DerInner -> BytesInner
The crate-internal `DerInner` type is really a `Cow`-like type for
generic bytes. The nature of those bytes is indicated by the pub wrapper
type holding the `DerInner`.

With that in mind we can re-use it for other non-DER types (e.g. ECH
configs) but first need to rename the type to `BytesInner` since it is
no longer DER in all cases.